Implement consumer and producer modules for transaction processing
- Added `consumer.rs` to handle file-based transaction processing with a priority queue for LSN ordering.
- Introduced `producer.rs` to manage transaction commits and notifications to the consumer.
- Updated `transaction_manager.rs` to expose `process_transaction_file` as a public method for broader access.
- Created integration tests for graceful shutdown scenarios, ensuring LSN persistence and proper handling of incomplete transactions.
- Enhanced error handling and logging throughout the consumer and producer workflows.
